pyproject.toml
[project]
name = "AnimeList"
version = "0.1.2"
description = "AnimeList e um app com foco em animes"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Fábio Adler", email = "adlerluna@hotmail.com" }
]
dependencies = [
  "flet==0.28.3",
  "bs4",
  "pillow",
  "requests"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.flet.android]
adaptive_icon_background = "#FFFFFF"

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "com.Aiterne.AnimeList"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "AnimeList"

# company name to display in about app dialogs
company = "Aiterne"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2025 by Flet"

[tool.flet.app]
path = "src"

[tool.uv]
dev-dependencies = [
    "flet[all]==0.28.3",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.28.3"}

[tool.flet.flutter.pubspec.dependency_overrides]
webview_flutter_android = "4.10.1"